Add a NodeFQDNAnnotation sync option for NodePort services
This supports the use case to sync Kubernetes NodePort services into Consul
using the FQDN of the node instead of an explicit IP address. It leverages
a new `consul.hashicorp.com/node-fqdn` annotation that needs to be applied
to each Kubernetes node to enable this sync type.

It would be preferable to be able to pull this information from the node's
information that is stored in Kubernetes, but there is little to no
documentation about that aspect of the system and many of the useful fields
are read-only (i.e. only populated by automatic Kubernetes processes).